requirements.txt
Flask==3.0.0
pygame==2.5.2
matplotlib==3.7.3
seaborn==0.12.2
pandas==2.1.1
numpy==1.26.0
Pillow==10.0.1
pydub==0.25.1
simpleaudio==1.0.4
ffmpeg==1.4
pyaudio==0.2.13
Jinja2==3.1.2
ffmpeg-python==0.2.0